Use MobileNetV3.Large as the backbone
With only fine tuning (MobileNet.trainable = false), the precision is
pretty bad: the validation set loss is only 0.0022 which is worse than
our previously best result of 0.0011.

However, if we allow the full network to be trained, then the validation
set loss is lowered to 6.3e-4 after 500 epochs. Moreover, it seems that
we can further reduce it using more epochs.